pyproject.toml
[project]
name = "mhars"
version = "1.0.0"
description = "Multi-modal Hybrid Adaptive Response System for thermal management"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Navin"}]

dependencies = [
    "gymnasium>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "stable-baselines3[extra]>=2.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
llm = ["llama-cpp-python>=0.2.0"]
iot = ["paho-mqtt>=2.0.0"]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["setuptools>=60.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "mhars*",
    "stage1_simulation*",
    "stage2_ml*",
    "stage3_ai*",
    "stage4_hardware*",
    "stage5_adapter*",
    "benchmarks*",
]
